# Improve List<T>.AddRange performance for enumerables