AddRange is currently implemented as delegating to InsertRange, and InsertRange in turn has a more complicated inner loop as part of adding each item from a source enumerable into the list.  By just copying InsertRange's source into AddRange, deleting all the irrelevant stuff, and changing the Insert call to Add, throughput improves measurably.

|              Method |         Toolchain | Count |        Mean | Ratio |
|-------------------- |------------------ |------ |------------:|------:|
| AddRange_Enumerable | \main\corerun.exe |     2 |    30.77 ns |  1.00 |
| AddRange_Enumerable |   \pr\corerun.exe |     2 |    26.35 ns |  0.85 |
|                     |                   |       |             |       |
| AddRange_Enumerable | \main\corerun.exe |   256 | 1,584.66 ns |  1.00 |
| AddRange_Enumerable |   \pr\corerun.exe |   256 | 1,069.29 ns |  0.69 |
|                     |                   |       |             |       |
| AddRange_Enumerable | \main\corerun.exe |  1024 | 6,223.58 ns |  1.00 |
| AddRange_Enumerable |   \pr\corerun.exe |  1024 | 4,175.18 ns |  0.67 |

```C#
private List<int> _list = new List<int>();
private IEnumerable<int> _sourceEnumerable;

[Params(2, 256, 1024)]
public int Count { get; set; }

[GlobalSetup]
public void Setup()
{
    _sourceEnumerable = Enumerable.Range(0, Count);
}

[Benchmark]
public void AddRange_Enumerable()
{
    _list.Clear();
    _list.AddRange(_sourceEnumerable);
}
```