@tfenise 2022-09-22
The benchmark code makes me wonder, why `Enumerable.Range` doesn't return some type implementing `ICollection<int>`, so that `List<int>.InsertRange` may call `ICollection<int>.CopyTo`.