@stephentoub 2022-09-23
> The benchmark code makes me wonder, why Enumerable.Range doesn't return some type implementing ICollection<int>, so that List<int>.InsertRange may call ICollection<int>.CopyTo.

It _could_; the question would be what scenario does that really serve, other than messing up my benchmarks :) It'd be more interesting I think to look at whether any of the other internal enumerable implementations in LINQ could/should implement `ICollection<T>`.