@stephentoub 2022-09-23
> t'd be more interesting I think to look at whether any of the other internal enumerable implementations in LINQ could/should implement ICollection<T>

I don't think it works.  The most valuable ones would be on the iterators used for things like Select (it's really rare, for example, to see an Enumerable.Range used in production where the result is consumed directly; it's much more common to see a Select or the like off of it).  The problem, however, is we'd need to run the selectors as part of `ICollection<T>.Count`, just as we do for `Enumerable.Count()`, and that means if someone did a fairly common pattern of using the Count to presize an output array and then CopyTo'ing into that array, it'd result in the selectors running twice.