@tfenise 2022-09-23
I want to comment on #1530, but I can't because that conversation "has been locked as resolved and limited to collaborators", even though that issue has been reopened, so I'll post here (please hide this as off-topic after that conversation is unlocked). I just want to note one thing. Adding the proposed overload `public void AddRange(ReadOnlySpan<T> collection);` (as well as `InsertRange`) would be a source code breaking change, because there would be an ambiguity when passing a `T[]` as the argument `collection`. Potential solutions include introducing another overload taking a `T[]` as parameter, or adding the proposed overload as an extension method, or maybe renaming the proposed overload (but we may want an argument of type `T[]` to go through the overload with parameter `ReadOnlySpan<T>` instead of `IEnumerable<T>`).