@stephentoub 2022-09-23
> I want to comment on https://github.com/dotnet/runtime/issues/1530, but I can't because that conversation "has been locked as resolved and limited to collaborators", even though that issue has been reopened,

I've unlocked it. Thanks.